fix score 0

+++ nodejs-express-sequelize-postgresql/dbInit/1init.sql
@@ -65,7 +65,7 @@ CREATE TABLE user_results (
     user_id int  NOT NULL,
     test_history_id int  NOT NULL,
     answers varchar(10000)  NOT NULL,
-    score int  NOT NULL,
+    score DECIMAL(10, 2)  NOT NULL,
     CONSTRAINT user_results_pk  PRIMARY KEY (id)
 );
